restarted building because i noticed my chat gpt "template" was just completed code and not what i wanted
