+++ main.py
@@ -1,21 +1,70 @@
-from perceptron import Perceptron
+import numpy as np
 
-def main():
-    # Example training data for AND logic gate
-    training_inputs = [
-        [0, 0],
-        [0, 1],
-        [1, 0],
-        [1, 1]
-    ]
-    labels = [0, 0, 0, 1]
+class Perceptron:
+    def __init__(self, learning_rate=0.01, n_iterations=1000):
+        """
+        Initialize the Perceptron with learning rate and number of iterations.
+        
+        Parameters:
+        learning_rate (float): The step size for weight updates.
+        n_iterations (int): Number of times to iterate over the training data.
+        
+        Initialize weights and bias to None.
+        """
+        # Initialize learning rate, number of iterations, weights, and bias here.
+        pass
 
-    perceptron = Perceptron(input_size=2)
-    perceptron.train(training_inputs, labels)
+    def _unit_step(self, x):
+        """
+        Unit step function for classification.
+        
+        Parameters:
+        x (float): The linear combination (score).
+        
+        Returns:
+        int: 1 if score >= 0, -1 otherwise.
+        
+        This function will determine the predicted label based on the score.
+        """
+        # Implement the step function here (return 1 or -1).
+        pass
 
-    # Test the perceptron
-    inputs = [1, 1]
-    print(f"Prediction for input {inputs}: {perceptron.predict(inputs)}")
+    def fit(self, X, y):
+        """
+        Train the Perceptron model by adjusting weights and bias.
+        
+        Parameters:
+        X (numpy.ndarray): Training data, shape (n_samples, n_features).
+        y (numpy.ndarray): True labels, shape (n_samples,).
+        
+        For each sample, calculate the score (dot product of features and weights),
+        make a prediction, and update the weights and bias if the prediction is incorrect.
+        """
+        # Initialize weights to zero and bias to 0.
+        # Implement the training loop over iterations and samples.
+        pass
 
+    def predict(self, X):
+        """
+        Make predictions using the trained Perceptron model.
+        
+        Parameters:
+        X (numpy.ndarray): Data to predict, shape (n_samples, n_features).
+        
+        Returns:
+        numpy.ndarray: Predicted labels (1 or -1).
+        
+        Compute the linear combination (score) and apply the unit step function to classify.
+        """
+        # Calculate the linear output using the learned weights and bias.
+        # Return predictions by applying the step function.
+        pass
+
+# Example usage (add real dataset in place of the dummy data):
 if __name__ == "__main__":
-    main()
+    
+    # Define the dataset (X: features, y: labels)
+    # Initialize the perceptron with learning rate and iterations
+    # Train the model using the `fit` function
+    # Use the `predict` function to classify new data
+    pass
